refactor: migrate utils function
